requirement.txt
colorama==0.4.6
coloredlogs==15.0.1
flatbuffers==23.5.26
humanfriendly==10.0
mpmath==1.3.0
numpy==1.21.6
onnxruntime==1.14.1
opencv-python==4.8.0.76
packaging==23.1
Pillow==9.5.0
protobuf==4.24.1
pyreadline==2.1
sympy==1.10.1
tqdm==4.66.1
